Database Test Cases done

+++ DataGenerator/src/dao/DatabaseDao.java
@@ -10,7 +10,7 @@ import exceptions.ReadEntityException;
 public interface DatabaseDao {
 	public void saveDatabse(Databasedetail databse) throws PersistException;
 
-	public Databasedetail getDatabaseByid(Long id) throws ReadEntityException;
+	public Databasedetail getDatabaseByid(Integer id) throws ReadEntityException;
 
 	public List<Databasedetail> getAllDatabaseinDB() throws ReadEntityException;
